+++ src/components/common/back-link-bar.tsx
@@ -5,11 +5,12 @@ import styles from '@/styles/components/common/back-link.module.scss';
 interface BackLinkBarProps {
     link: string; // TODO: various pages links
     label: string;
+    onlyMarginBottom?: boolean;
 }
 
-export function BackLinkBar({ link, label }: BackLinkBarProps) {
+export function BackLinkBar({ link, label, onlyMarginBottom }: BackLinkBarProps) {
     return (
-        <div className={styles['back-link-bar']}>
+        <div className={styles['back-link-bar']} data-only-margin-bottom={onlyMarginBottom}>
             <Link href={link}>
                 <ArrowCircleLeftIcon />{label}
             </Link>