+++ src/types/page.types.ts
@@ -1,0 +1,4 @@
+export interface PageProps<ParamsKeyList extends string> {
+    params: Record<ParamsKeyList, string>;
+    searchParams?: { [key: string]: string | string[] | undefined };
+}